Fix PlaylistScroller controller visibility bug
